+++ constants/networks.ts
@@ -1,6 +1,12 @@
 export const ETH = 'ethereum';
 export const BNB = 'bnb';
 export const BNB_SANDBOX = 'bnb_sandbox';
+export const XDC_SANDBOX = 'xdc_sandbox';
+export const OKX_SANDBOX = 'okx_sandbox';
+export const ARTHERA_SANDBOX = 'arthera_sandbox';
+export const ASTAR_SANDBOX = 'astar_sandbox';
+export const MELD_SANDBOX = 'meld_sandbox';
+export const ROOTSTOCK_SANDBOX = 'rootstock_sandbox';
 
 export const LOCALHOST = 'localhost';
 export const HARDHAT = 'hardhat';
@@ -16,18 +22,22 @@ export const BNB_STAGING = 'bnb_staging';
 export const ETH_STAGING = 'eth_staging';
 export const POLYGON_STAGING = 'polygon_staging';
 
+export const ZK_LINK_NOVA_STAGING = 'zk_link_nova_staging';
+
 export const ARBITRUM_SANDBOX = 'arbitrum_sandbox';
 export const AVALANCHE_SANDBOX = 'avalanche_sandbox';
 export const POLYGON_SANDBOX = 'polygon_sandbox';
 export const ETH_SANDBOX = 'eth_sandbox';
+export const ZK_LINK_NOVA_SANDBOX = 'zk_link_nova_sandbox';
 
+export const ZK_LINK_NOVA_PRODUCTION = 'zk_link_nova_production';
 export const ARBITRUM_PRODUCTION = 'arbitrum_production';
 export const AVALANCHE_PRODUCTION = 'avalanche_production';
 export const ETH_PRODUCTION = 'eth_production';
 export const POLYGON_PRODUCTION = 'polygon_production';
 export const BNB_PRODUCTION = 'bnb_production';
 export const LINEA_PRODUCTION = 'linea_production';
-
+export const ROOTSTOCK_PRODUCTION = 'rootstock_production';
 
 export const MASTER_CHAIN_NAME = BNB;
 
@@ -45,8 +55,7 @@ export const isMasterChain = (chainId: number | string): boolean => {
     return FORKING_ENV == MASTER_CHAIN_NAME;
   }
 
-  const masterChain = [LOCALHOST_ID, BNB_ID, BNB_STAGING_ID, FORKED_BNB_ID]
-    .includes(parseInt(`${chainId}`, 10));
+  const masterChain = [LOCALHOST_ID, BNB_ID, BNB_STAGING_ID, FORKED_BNB_ID].includes(parseInt(`${chainId}`, 10));
 
   console.log(`isMasterChain(${chainId})?`, masterChain);
   return masterChain;