+++ hardhat.config.ts
@@ -6,8 +6,10 @@ import 'hardhat-deploy';
 import 'hardhat-deploy-ethers';
 import 'hardhat-gas-reporter';
 import '@nomiclabs/hardhat-solhint';
-import '@nomiclabs/hardhat-etherscan';
+import '@nomicfoundation/hardhat-verify';
 import 'solidity-coverage';
+import '@matterlabs/hardhat-zksync-solc';
+import '@matterlabs/hardhat-zksync-verify';
 
 // there is undefined issue in this repo, ts-node is ignoring flag TS_NODE_TRANSPILE_ONLY=1 and throw errors
 // on missing typechain/
@@ -19,12 +21,34 @@ import {HardhatNetworkForkingUserConfig, HardhatUserConfig} from 'hardhat/types'
 import {
   ARBITRUM_PRODUCTION,
   ARBITRUM_SANDBOX,
-  ARBITRUM_STAGING, AVALANCHE_PRODUCTION, AVALANCHE_SANDBOX,
-  AVALANCHE_STAGING, BASE_PRODUCTION, BASE_STAGING,
-  BNB, BNB_PRODUCTION, BNB_SANDBOX,
-  BNB_STAGING, ETH, ETH_PRODUCTION, ETH_SANDBOX,
-  ETH_STAGING, LINEA_PRODUCTION, LINEA_SANDBOX, LINEA_STAGING,
-  LOCALHOST, POLYGON_PRODUCTION, POLYGON_SANDBOX, POLYGON_STAGING
+  ARBITRUM_STAGING,
+  ARTHERA_SANDBOX,
+  ASTAR_SANDBOX,
+  AVALANCHE_PRODUCTION,
+  AVALANCHE_SANDBOX,
+  AVALANCHE_STAGING,
+  BASE_PRODUCTION,
+  BASE_STAGING,
+  BNB,
+  BNB_PRODUCTION,
+  BNB_SANDBOX,
+  BNB_STAGING,
+  ETH,
+  ETH_PRODUCTION,
+  ETH_SANDBOX,
+  ETH_STAGING,
+  LINEA_PRODUCTION,
+  LINEA_SANDBOX,
+  LINEA_STAGING,
+  LOCALHOST,
+  MELD_SANDBOX,
+  OKX_SANDBOX,
+  POLYGON_PRODUCTION,
+  POLYGON_SANDBOX,
+  POLYGON_STAGING,
+  ROOTSTOCK_PRODUCTION,
+  ROOTSTOCK_SANDBOX,
+  XDC_SANDBOX, ZK_LINK_NOVA_PRODUCTION, ZK_LINK_NOVA_SANDBOX, ZK_LINK_NOVA_STAGING
 } from './constants/networks';
 import {getPrivteKeys, PROD_PK} from './constants/pk';
 import {forkingChainId, getProviderData} from './constants/providers';
@@ -72,7 +96,7 @@ const apiKey = (): string | Record<string, string> => {
     'lineatestnet': LINEASCAN_API,
     'linea': LINEASCAN_API,
     'base-goerli': 'PLACEHOLDER_STRING',
-    'base-mainnet': BASESCAN_API
+    'base-mainnet': BASESCAN_API,
   };
 };
 
@@ -242,6 +266,72 @@ const config: HardhatUserConfig = {
       chainId: getProviderData(BNB_SANDBOX).chainId,
       gasPrice: gwei(10)
     },
+    xdc_sandbox: {
+      url: getProviderData(XDC_SANDBOX).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(XDC_SANDBOX).chainId,
+      gasPrice: 'auto'
+    },
+    okx_sandbox: {
+      url: getProviderData(OKX_SANDBOX).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(OKX_SANDBOX).chainId,
+      gasPrice: 'auto'
+    },
+    astar_sandbox: {
+      url: getProviderData(ASTAR_SANDBOX).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(ASTAR_SANDBOX).chainId,
+      gasPrice: 'auto'
+    },
+    arthera_sandbox: {
+      url: getProviderData(ARTHERA_SANDBOX).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(ARTHERA_SANDBOX).chainId,
+      gasPrice: 'auto'
+    },
+    meld_sandbox: {
+      url: getProviderData(MELD_SANDBOX).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(MELD_SANDBOX).chainId,
+      gasPrice: 'auto'
+    },
+    rootstock_sandbox: {
+      url: getProviderData(ROOTSTOCK_SANDBOX).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(ROOTSTOCK_SANDBOX).chainId,
+      gasPrice: 'auto'
+    },
+    zk_link_nova_staging: {
+      url: getProviderData(ZK_LINK_NOVA_STAGING).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(ZK_LINK_NOVA_STAGING).chainId,
+      gasPrice: 'auto',
+      ethNetwork: 'sepolia', // The Ethereum Web3 RPC URL, or the identifier of the network (e.g. `mainnet`, `sepolia`)
+      // Verification endpoint for Sepolia
+      verifyURL: 'https://explorer.sepolia.era.zksync.dev/contract_verification',
+      zksync: true, // enables zksolc compiler
+    },
+    zk_link_nova_sandbox: {
+      url: getProviderData(ZK_LINK_NOVA_SANDBOX).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(ZK_LINK_NOVA_SANDBOX).chainId,
+      gasPrice: 'auto',
+      ethNetwork: 'sepolia', // The Ethereum Web3 RPC URL, or the identifier of the network (e.g. `mainnet`, `sepolia`)
+      // Verification endpoint for Sepolia
+      verifyURL: 'https://explorer.sepolia.era.zksync.dev/contract_verification',
+      zksync: true, // enables zksolc compiler
+    },
+    zk_link_nova_productin: {
+      url: getProviderData(ZK_LINK_NOVA_PRODUCTION).url,
+      accounts: getPrivteKeys(LOCALHOST),
+      chainId: getProviderData(ZK_LINK_NOVA_PRODUCTION).chainId,
+      gasPrice: 'auto',
+      ethNetwork: 'mainnet', // The Ethereum Web3 RPC URL, or the identifier of the network (e.g. `mainnet`, `sepolia`)
+      // Verification endpoint for Sepolia
+      verifyURL: 'https://explorer.era.zksync.dev/contract_verification',
+      zksync: true, // enables zksolc compiler
+    },
     arbitrum_production: {
       url: getProviderData(ARBITRUM_PRODUCTION).url,
       accounts: getPrivteKeys(PROD_PK),
@@ -287,6 +377,12 @@ const config: HardhatUserConfig = {
       url: getProviderData(BASE_PRODUCTION).url,
       accounts: getPrivteKeys(PROD_PK),
       chainId: getProviderData(BASE_PRODUCTION).chainId,
+      live: true
+    },
+    rootstock_production: {
+      url: getProviderData(ROOTSTOCK_PRODUCTION).url,
+      accounts: getPrivteKeys(PROD_PK),
+      chainId: getProviderData(ROOTSTOCK_PRODUCTION).chainId,
       live: true
     },
     docker: {
@@ -364,6 +460,27 @@ const config: HardhatUserConfig = {
         },
       }
     ]
+  },
+  zksolc: {
+    version: '1.4.0', // optional eg 'latest'
+    settings: {
+      // compilerPath: 'zksolc',  // optional. Can be used if compiler is located in a specific folder
+      libraries:{}, // optional. References to non-inlinable libraries
+      missingLibrariesPath: './.zksolc-libraries-cache/missingLibraryDependencies.json', // optional. cache
+      isSystem: false, // optional.  Enables Yul instructions available only for zkSync system contracts and libraries
+      forceEvmla: false, // optional. Falls back to EVM legacy assembly if there is a bug with Yul
+      optimizer: {
+        enabled: true, // optional. True by default
+        mode: '3', // optional. 3 by default, z to optimize bytecode size
+        // optional. Try to recompile with optimizer mode "z" if the bytecode is too large
+        fallback_to_optimizing_for_size: false,
+      },
+      experimental: {
+        dockerImage: '', // deprecated
+        tag: ''   // deprecated
+      },
+      contractsToCompile: [] //optional. Compile only specific contracts
+    }
   },
   namedAccounts: {
     deployer: 0,